F7 Inav, make things work for now

+++ src/main/drivers/serial_uart_hal.c
@@ -283,7 +283,7 @@ uint32_t uartTotalRxBytesWaiting(serialPort_t *instance)
     }
 }
 
-uint32_t uartTotalTxBytesFree(serialPort_t *instance)
+uint8_t uartTotalTxBytesFree(serialPort_t *instance)
 {
     uartPort_t *s = (uartPort_t*)instance;
